export dune queries func
